vertmodel cosf->std::cos for #52

+++ src/engine/model/vertmodel.h
@@ -638,7 +638,7 @@ struct vertcommands : modelcommands<MDL, struct MDL::vertmesh>
         {
             mdl.disablepitch();
         }
-        mdl.meshes = MDL::loading->sharemeshes(path(filename), *smooth > 0 ? cosf(std::clamp(*smooth, 0.0f, 180.0f)*RAD) : 2);
+        mdl.meshes = MDL::loading->sharemeshes(path(filename), *smooth > 0 ? std::cos(std::clamp(*smooth, 0.0f, 180.0f)*RAD) : 2);
         if(!mdl.meshes)
         {
             conoutf("could not load %s", filename);
@@ -657,11 +657,11 @@ struct vertcommands : modelcommands<MDL, struct MDL::vertmesh>
             return;
         }
         part &mdl = *static_cast<part *>(MDL::loading->parts.last());
-        float cx = *rx ? cosf(*rx/2*RAD) : 1,
+        float cx = *rx ? std::cos(*rx/2*RAD) : 1,
               sx = *rx ? sinf(*rx/2*RAD) : 0,
-              cy = *ry ? cosf(*ry/2*RAD) : 1,
+              cy = *ry ? std::cos(*ry/2*RAD) : 1,
               sy = *ry ? sinf(*ry/2*RAD) : 0,
-              cz = *rz ? cosf(*rz/2*RAD) : 1,
+              cz = *rz ? std::cos(*rz/2*RAD) : 1,
               sz = *rz ? sinf(*rz/2*RAD) : 0;
         //matrix m created from (matrix3 created from quat) + (vec) appended afterwards
         matrix4x3 m(static_cast<matrix3>(quat(sx*cy*cz - cx*sy*sz,
